Introducing the risk manager.
It will monitor user margins to make sure they aren't being naughty.
I am progressively migrating the margin function out of the accountant so it can be shared.
Also introducing a hub for safe price publishing.
I will refactor so that the webserver and the accountant use this device. There are related config file changes.