Fix order management. If the internal order got removed right after placeorder came back then there would be a period between that time and when the order showed up in the feed that we don't have a record of that order that we have outstanding. Now we just keep track of the order id to internal id map and then when the order comes in through the feed, we delete the internal order
this is better than just assigning the right id to the list of orders in onPlaceOrder, because we don't know what order the feed message or the onplaceorder return will happen in. If the feed message comes in first and the quantity was updated, and then the onplaceorder returns, then we might clobber thef act that some things got filled.