closes #647 closes #648 closes #649
making this work properly with the tests is fairly involved. Better to evaluate and solve
the performance issue in place_order and deal with what happens in the accountant
if place_order times out and we aren't able to mark the order dispatched even though
it ventually reaches the book.

Revert "closes #646"

This reverts commit 5adb27b71e911e1a14d6245a676ae9a4e741914d.

Conflicts:
	github